seperate non html specific graph data generation from htmlhelper
