+++ LuckParser/LuckParser.csproj
@@ -78,6 +78,7 @@
     <Compile Include="Controllers\Controller1.cs" />
     <Compile Include="Controllers\CSVBuilder.cs" />
     <Compile Include="Controllers\FlashWindow.cs" />
+    <Compile Include="Controllers\GraphHelper.cs" />
     <Compile Include="Controllers\GW2APIController.cs" />
     <Compile Include="Controllers\HTMLBuilder.cs" />
     <Compile Include="Controllers\HTMLHelper.cs" />
